+++ examples/agentic-editing/CHANGELOG.md
@@ -1,5 +1,12 @@
 # agentic-editing-example
 
+## 0.0.1-alpha.14
+
+### Patch Changes
+
+- Updated dependencies [[`d9c79e8`](https://github.com/mattppal/mcut/commit/d9c79e82e236119e64192b02ece2529e1cac6a38)]:
+  - @mcut/timeline@0.1.0-alpha.15
+
 ## 0.0.1-alpha.13
 
 ### Patch Changes