+++ packages/transcription-ai-sdk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mcut/transcription-ai-sdk
 
+## 1.0.0-alpha.16
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mcut/transcription@0.1.0-alpha.16
+
 ## 1.0.0-alpha.15
 
 ### Patch Changes